fix the issue that tmp files are not all deleted if from restart

+++ grad_defn.py
@@ -110,11 +110,13 @@ def grad_defn(process, u_pert, t, epsilon, restart=False, iter0=None):
     mpi_comm.Allreduce(g_local, g_global, op=process.mpi.SUM)
     logging.debug("Rank {}: Gradients gathered".format(mpi_rank))
 
-    # At these stage, all ranks have computed/loaded the gradients, so we can delete the tmp files
-    for index in my_indices:
-        tmp_fn = "{}/tmp_grad_defn_iter_{}_index_{}_{}_{}.npy".format(mpi_root_dir, iter0,
-                                                                      index[0], index[1], index[2])
-        pathlib.Path(tmp_fn).unlink()
+    # At these stage, all ranks have computed/loaded the gradients, so we can delete the tmp files for iter0
+    if mpi_rank == 0:
+        for index in list(np.ndindex(shape)):
+            tmp_fn = "{}/tmp_grad_defn_iter_{}_index_{}_{}_{}.npy".format(mpi_root_dir, iter0,
+                                                                          index[0], index[1], index[2])
+            if pathlib.Path(tmp_fn).exists():
+                pathlib.Path(tmp_fn).unlink()
 
     return g_global
